requirements.txt
python-dotenv~=1.0.1
dataclasses~=0.6
PyJWT~=2.10.1
bcrypt~=4.2.1
google~=3.0.0
SQLAlchemy~=2.0.36
pydantic~=2.10.0